requirements.txt
huggingface_hub==0.25.2
pandas==2.2.3
scikit-learn==1.6.1
nltk==3.9.1
matplotlib==3.10.0
Flask==3.1.0
Flask-Cors==5.0.0
black==24.10.0
python-dotenv==1.0.1
evaluate==0.4.3
langchain-openai==0.2.10
langchain-cerebras==0.3.0
langchain-ollama==0.2.0
langchain==0.3.13
langchain-community==0.3.13
jupyter==1.1.1
ipywidgets==8.1.5
seaborn==0.13.2
numpy==1.26.4
langgraph==0.2.60
transformers==4.48.0
torch==2.2.2
psutil==6.1.1
google-genai==0.6.0
